--- orchestrator/pandora-chain/log.go
@@ -1,1 +1,0 @@
-package pandora_chain
