--- orchestrator/pandora-chain/service_test.go
@@ -1,1 +1,0 @@
-package pandora_chain
