--- orchestrator/vanguard-chain/sevice_test.go
@@ -1,1 +1,0 @@
-package vanguard_chain
