+++ orchestrator/vanguardchain/client/client_test.go
