+++ orchestrator/vanguardchain/subscriber/log.go
